+++ src/types.js
@@ -1287,27 +1287,29 @@ class ResourceNode {
    *  Constructs a instance for the given node ("data") of a resource.  If the
    *  data is the top-level node of a resouce, the path and type parameters will
    *  be ignored in favor of the resource's resourceType field.
-   * @param {*} data the node's data or value (which might be an object with
+   * @param {*} data - the node's data or value (which might be an object with
    *  sub-nodes, an array, or FHIR data type)
-   * @param {string} path the node's path in the resource (e.g. Patient.name).
+   * @param {ResourceNode} parentResNode - parent ResourceNode.
+   * @param {string} path - the node's path in the resource (e.g. Patient.name).
    *  If the data's type can be determined from data, that will take precedence
    *  over this parameter.
-   * @param {*} _data additional data stored in a property named with "_"
+   * @param {*} _data - additional data stored in a property named with "_"
    *  prepended, see https://www.hl7.org/fhir/element.html#json for details.
-   * @param {string} fhirNodeDataType FHIR node data type, if the resource node
+   * @param {string} fhirNodeDataType - FHIR node data type, if the resource node
    *  is described in the FHIR model.
    */
-  constructor(data, path, _data, fhirNodeDataType) {
+  constructor(data, parentResNode, path, _data, fhirNodeDataType) {
     // If data is a resource (maybe a contained resource) reset the path
     // information to the resource type.
     if (data?.resourceType) {
       path = data.resourceType;
       fhirNodeDataType = data.resourceType;
     }
-    this.path = path;
+    this.parentResNode = parentResNode || null;
+    this.path = path || null;
     this.data = data;
     this._data = _data || {};
-    this.fhirNodeDataType = fhirNodeDataType;
+    this.fhirNodeDataType = fhirNodeDataType || null;
   }
 
   /**
@@ -1384,8 +1386,8 @@ class ResourceNode {
  *  given node is already a ResourceNode.  Takes the same arguments as the
  *  constructor for ResourceNode.
  */
-ResourceNode.makeResNode = function(data, path, _data, fhirNodeDataType = null) {
-  return (data instanceof ResourceNode) ? data : new ResourceNode(data, path, _data, fhirNodeDataType);
+ResourceNode.makeResNode = function(data, parentResNode, path, _data, fhirNodeDataType = null) {
+  return (data instanceof ResourceNode) ? data : new ResourceNode(data, parentResNode, path, _data, fhirNodeDataType);
 };
 
 // The set of available data types in the System namespace
@@ -1523,6 +1525,49 @@ TypeInfo.fromValue = function (value) {
 };
 
 /**
+ * Set of primitive data type names.
+ */
+const primitives = new Set();
+// IE11 probably doesn't support `new Set(iterable)`
+[
+  "instant",
+  "time",
+  "date",
+  "dateTime",
+  "base64Binary",
+  "decimal",
+  "integer64",
+  "boolean",
+  "string",
+  "code",
+  "markdown",
+  "id",
+  "integer",
+  "unsignedInt",
+  "positiveInt",
+  "uri",
+  "oid",
+  "uuid",
+  "canonical",
+  "url",
+  "Integer",
+  "Decimal",
+  "String",
+  "Date",
+  "DateTime",
+  "Time"
+].forEach(i => primitives.add(i));
+
+/**
+ * Checks whether the specified type information contains a primitive data type.
+ * @param {TypeInfo} typeInfo
+ * @return {boolean}
+ */
+TypeInfo.isPrimitive = function(typeInfo) {
+  return primitives.has(typeInfo.name);
+};
+
+/**
  * Basic "type()" function implementation
  * (see http://hl7.org/fhirpath/#reflection)
  * @param {Array<*>} coll - input collection
@@ -1573,17 +1618,19 @@ function asFn(coll, typeInfo) {
 }
 
 module.exports = {
-  FP_Type: FP_Type,
-  FP_TimeBase: FP_TimeBase,
-  FP_Date: FP_Date,
-  FP_DateTime: FP_DateTime,
-  FP_Instant: FP_Instant,
-  FP_Time: FP_Time,
-  FP_Quantity: FP_Quantity,
-  timeRE: timeRE,
-  dateTimeRE: dateTimeRE,
-  ResourceNode: ResourceNode,
-  TypeInfo: TypeInfo,
+  FP_Type,
+  FP_TimeBase,
+  FP_Date,
+  FP_DateTime,
+  FP_Instant,
+  FP_Time,
+  FP_Quantity,
+  timeRE,
+  dateTimeRE,
+  dateRE,
+  instantRE,
+  ResourceNode,
+  TypeInfo,
   typeFn,
   isFn,
   asFn